requirements.txt
streamlit==1.32.0
tensorflow
numpy
pandas
pillow